requirements.txt
torch<=2.3
torchaudio
modelscope
huggingface
huggingface_hub
funasr>=1.1.3
numpy<=1.26.4
gradio
fastapi>=0.111.1
PyQt5
Pillow
pydub
simpleaudio
pyaudio
webrtcvad
requests
zhipuai
